Add Talos II configuration files
The Raptor Computing Systems Talos II is a POWER9 design that was
derived from Romulus.

These are based on the raptor-v1.05 tag in their git repository:
https://git.raptorcs.com/git/talos-op-build/

Additional changes have been made in an attempt to align with upstream,
disable as many blobs as is currently possible, and remove references
to any changes that will not be upstreamed.

Additionally, users should note that unlike downstream, this firmware
build is *not* blob free, as that requires work upstream.

The configurations may not build functional firmware without future
changes to align with downstream. It is currently untested.

Signed-off-by: Alexander Lent <ppc64@xanderlent.com>